requirements.txt
streamlit
pytesseract
pdf2image
openai
Pillow
python-dotenv
PyMuPDF